move withNextRouter to __tests__/utils folder
