+++ stories/pages/Lesson.stories.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import Lesson from '../../pages/curriculum/[lesson]'
 import { MockedProvider } from '@apollo/client/testing'
 import GET_APP from '../../graphql/queries/getApp'
-import { withTestRouter } from '../../testUtil/withNextRouter'
+import { withTestRouter } from '../../__tests__/utils/withNextRouter'
 import dummyLessonData from '../../__dummy__/lessonData'
 import dummySessionData from '../../__dummy__/sessionData'
 import dummyAlertData from '../../__dummy__/alertData'
