+++ stories/pages/Login.stories.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import LoginPage, { Login } from '../../pages/login'
 import { MockedProvider } from '@apollo/client/testing'
 import GET_APP from '../../graphql/queries/getApp'
-import { withTestRouter } from '../../testUtil/withNextRouter'
+import { withTestRouter } from '../../__tests__/utils/withNextRouter'
 
 export default {
   component: Login,
